Crate display_interface

A generic display interface

This crate contains an error type and traits to implement for bus interface drivers drivers to be consumed by display drivers. It abstracts over the different communication methods available to drive a display and allows a driver writer to focus on driving the display itself and only have to implement a single interface.

Modules

prelude

Enums

DataFormat

DI specific data format wrapper around slices of various widths Display drivers need to implement non-trivial conversions (e.g. with padding) as the hardware requires.

DisplayError

A ubiquitous error type for all kinds of problems which could happen when communicating with a display

Traits

AsyncWriteOnlyDataCommand

This trait implements a write-only interface for a display which has separate data and command modes. It is the responsibility of implementations to activate the correct mode in their implementation when corresponding method is called.

WriteOnlyDataCommand

This trait implements a write-only interface for a display which has separate data and command modes. It is the responsibility of implementations to activate the correct mode in their implementation when corresponding method is called.